Fix Makefile and fmod declaration

+++ src/sysdolphin/baselib/aobj.h
@@ -52,7 +52,7 @@ void HSD_AObjInvokeCallBacks(void);
 void HSD_AObjReqAnim(HSD_AObj* aobj, f32 frame);
 void HSD_AObjStopAnim(HSD_AObj* aobj, void* obj, void (*func)());
 void HSD_AObjInterpretAnim(HSD_AObj* aobj, void* obj, void (*update_func)());
-f32 fmod(f64 x, f64 y);
+float fmod(float x, float y);
 HSD_AObj* HSD_AObjLoadDesc(HSD_AObjDesc* aobjdesc);
 void HSD_AObjRemove(HSD_AObj* aobj);
 HSD_AObj* HSD_AObjAlloc(void);
